chore(arch): refacto models
