+++ app/Filament/Resources/UserResource.php
@@ -3,7 +3,7 @@
 namespace App\Filament\Resources;
 
 use App\Filament\Resources\UserResource\Pages;
-use App\Models\User;
+use App\Models\User\User;
 use Filament\Forms;
 use Filament\Forms\Form;
 use Filament\Resources\Resource;
