+++ app/Models/Marketplace/Product.php
@@ -1,22 +1,22 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Marketplace;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Wallet extends Model
+class Product extends Model
 {
     use SoftDeletes;
 
     protected $fillable = [
-        'user_id',
-        'balance',
+        'product_type_id',
+        'price',
     ];
 
-    public function user(): BelongsTo
+    public function productType(): BelongsTo
     {
-        return $this->belongsTo(User::class, 'user_id');
+        return $this->belongsTo(Type::class, 'product_type_id');
     }
 }
