+++ app/Models/Pomodoro.php
@@ -2,11 +2,12 @@
 
 namespace App\Models;
 
+use App\Models\User\User;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class PomodoroSession extends Model
+class Pomodoro extends Model
 {
     use SoftDeletes;
