+++ app/Policies/UserPolicy.php
@@ -2,7 +2,7 @@
 
 namespace App\Policies;
 
-use App\Models\User;
+use App\Models\User\User;
 
 class UserPolicy
 {
