+++ database/seeders/TransactionLedgerSeeder.php
@@ -2,9 +2,9 @@
 
 namespace Database\Seeders;
 
-use App\Models\Transaction;
-use App\Models\TransactionLedger;
-use App\Models\User;
+use App\Models\User\User;
+use App\Models\Wallet\Transaction;
+use App\Models\Wallet\TransactionLedger;
 use Illuminate\Database\Seeder;
 
 class TransactionLedgerSeeder extends Seeder
@@ -17,15 +17,17 @@ class TransactionLedgerSeeder extends Seeder
         $transaction = Transaction::first();
         $users = User::all();
 
-        if($transaction)
-            foreach($users as $user)
+        if ($transaction) {
+            foreach ($users as $user) {
                 TransactionLedger::create([
                     'transaction_id' => $transaction->id,
                     'user_id' => $user->id,
                     'type' => 'credit',
                     'amount' => 25.00,
                 ]);
-        else
+            }
+        } else {
             throw new \Exception('No transaction found to link with TransactionLedger');
+        }
     }
 }
