+++ routes/console.php
@@ -4,5 +4,5 @@ use Illuminate\Foundation\Inspiring;
 use Illuminate\Support\Facades\Artisan;
 
 Artisan::command('inspire', function () {
-    echo Inspiring::quote() . PHP_EOL;
+    echo Inspiring::quote().PHP_EOL;
 })->purpose('Display an inspiring quote')->hourly();
